+++ src/Bootstrap/ContainerCreatedObserver.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Cspray\AnnotatedContainer\Bootstrap;
+
+use Cspray\AnnotatedContainer\AnnotatedContainer;
+use Cspray\AnnotatedContainer\Definition\ContainerDefinition;
+use Cspray\AnnotatedContainer\Profiles\ActiveProfiles;
+
+interface ContainerCreatedObserver {
+
+    public function notifyContainerCreated(ActiveProfiles $activeProfiles, ContainerDefinition $containerDefinition, AnnotatedContainer $container) : void;
+
+}