+++ src/Exception/InvalidBootstrapConfiguration.php
@@ -2,7 +2,10 @@
 
 namespace Cspray\AnnotatedContainer\Exception;
 
+use Cspray\AnnotatedContainer\Bootstrap\ContainerCreatedObserver;
 use Cspray\AnnotatedContainer\Bootstrap\Observer;
+use Cspray\AnnotatedContainer\Bootstrap\PostAnalysisObserver;
+use Cspray\AnnotatedContainer\Bootstrap\PreAnalysisObserver;
 use Cspray\AnnotatedContainer\StaticAnalysis\DefinitionProvider;
 use Cspray\AnnotatedContainer\ContainerFactory\ParameterStore;
 
@@ -31,8 +34,10 @@ final class InvalidBootstrapConfiguration extends Exception {
 
     public static function fromConfiguredObserverWrongType() : self {
         return new self(sprintf(
-            'All entries in observers must be classes that implement %s',
-            Observer::class
+            'All entries in observers must be classes that implement %s, %s, or %s',
+            PreAnalysisObserver::class,
+            PostAnalysisObserver::class,
+            ContainerCreatedObserver::class
         ));
     }