+++ test/Unit/Helper/StubBootstrapObserver.php
@@ -3,27 +3,25 @@
 namespace Cspray\AnnotatedContainer\Unit\Helper;
 
 use Cspray\AnnotatedContainer\AnnotatedContainer;
-use Cspray\AnnotatedContainer\Bootstrap\Observer;
+use Cspray\AnnotatedContainer\Bootstrap\ContainerCreatedObserver;
+use Cspray\AnnotatedContainer\Bootstrap\PostAnalysisObserver;
+use Cspray\AnnotatedContainer\Bootstrap\PreAnalysisObserver;
 use Cspray\AnnotatedContainer\Definition\ContainerDefinition;
 use Cspray\AnnotatedContainer\Profiles\ActiveProfiles;
 
-class StubBootstrapObserver implements Observer {
+class StubBootstrapObserver implements PreAnalysisObserver, PostAnalysisObserver, ContainerCreatedObserver {
 
     private array $invokedMethods = [];
 
-    public function beforeCompilation(ActiveProfiles $activeProfiles) : void {
+    public function notifyPreAnalysis(ActiveProfiles $activeProfiles) : void {
         $this->invokedMethods[] = [__METHOD__];
     }
 
-    public function afterCompilation(ActiveProfiles $activeProfiles, ContainerDefinition $containerDefinition) : void {
+    public function notifyPostAnalysis(ActiveProfiles $activeProfiles, ContainerDefinition $containerDefinition) : void {
         $this->invokedMethods[] = [__METHOD__];
     }
 
-    public function beforeContainerCreation(ActiveProfiles $activeProfiles, ContainerDefinition $containerDefinition) : void {
-        $this->invokedMethods[] = [__METHOD__];
-    }
-
-    public function afterContainerCreation(ActiveProfiles $activeProfiles, ContainerDefinition $containerDefinition, AnnotatedContainer $container) : void {
+    public function notifyContainerCreated(ActiveProfiles $activeProfiles, ContainerDefinition $containerDefinition, AnnotatedContainer $container) : void {
         $this->invokedMethods[] = [__METHOD__];
     }